+++ lib/canmove.cpp
@@ -2,11 +2,11 @@
 using namespace std;
 using ll = long long;
 
-vector<ll> dx = {0, 0, -1, 1};
-vector<ll> dy = {-1, 1, 0, 0};
-ll H, W;
+vector<int> dx = {0, 0, -1, 1};
+vector<int> dy = {-1, 1, 0, 0};
+int H, W;
 
-bool canmove(ll x, ll y) {
+bool canmove(int x, int y) {
     if (0 <= x && x < H && 0 <= y && y < W) return true;
     else return false;
 }
